Small refactor; include reducer
