+++ _inc/client/state/mobile/index.js
@@ -1,0 +1,9 @@
+/**
+ * Internal dependencies
+ */
+import * as reducer from './reducer';
+import * as actions from './actions';
+
+const all = { ...reducer, ...actions };
+
+export default all;
