+++ _inc/client/state/reducer.js
@@ -23,6 +23,7 @@ import { reducer as search } from 'state/search/reducer';
 import { reducer as devCard } from 'state/dev-version/reducer';
 import { reducer as publicize } from 'state/publicize/reducer';
 import { reducer as siteVerify } from 'state/site-verify/reducer';
+import { reducer as mobile } from 'state/mobile/reducer';
 
 const jetpackReducer = combineReducers( {
 	initialState,
@@ -40,6 +41,7 @@ const jetpackReducer = combineReducers( {
 	devCard,
 	publicize,
 	siteVerify,
+	mobile,
 } );
 
 export default combineReducers( {
